+++ packages/react-components/lib/tasks/create-task.tsx
@@ -5,9 +5,7 @@
 
 import UpdateIcon from '@mui/icons-material/Create';
 import DeleteIcon from '@mui/icons-material/Delete';
-import PlaceOutlined from '@mui/icons-material/PlaceOutlined';
 import {
-  Autocomplete,
   Box,
   Button,
   Checkbox,
@@ -25,7 +23,6 @@ import {
   IconButton,
   List,
   ListItem,
-  ListItemIcon,
   ListItemSecondaryAction,
   ListItemText,
   MenuItem,
@@ -42,7 +39,6 @@ import type { TaskFavoritePydantic as TaskFavorite, TaskRequest } from 'api-clie
 import React from 'react';
 import { Loading } from '..';
 import { ConfirmationDialog, ConfirmationDialogProps } from '../confirmation-dialog';
-import { PositiveIntField } from '../form-inputs';
 import {
   defaultDeliveryCustomTaskDescription,
   defaultDeliveryTaskDescription,
@@ -51,12 +47,7 @@ import {
   DeliveryTaskDescription,
   DeliveryTaskForm,
 } from './descriptions/delivery_custom';
-
-// A bunch of manually defined descriptions to avoid using `any`.
-export interface PatrolTaskDescription {
-  places: string[];
-  rounds: number;
-}
+import { defaultPatrolTask, PatrolTaskDescription, PatrolTaskForm } from './descriptions/patrol';
 
 type CustomComposeTaskDescription = string;
 
@@ -64,18 +55,6 @@ type TaskDescription =
   | DeliveryTaskDescription
   | DeliveryCustomTaskDescription
   | PatrolTaskDescription;
-
-const isPatrolTaskDescriptionValid = (taskDescription: PatrolTaskDescription): boolean => {
-  if (taskDescription.places.length === 0) {
-    return false;
-  }
-  for (const place of taskDescription.places) {
-    if (place.length === 0) {
-      return false;
-    }
-  }
-  return taskDescription.rounds > 0;
-};
 
 const isCustomTaskDescriptionValid = (taskDescription: string): boolean => {
   if (taskDescription.length === 0) {
@@ -118,122 +97,6 @@ const StyledDialog = styled((props: DialogProps) => <Dialog {...props} />)(({ th
     minWidth: 80,
   },
 }));
-
-interface PlaceListProps {
-  places: string[];
-  onClick(places_index: number): void;
-}
-
-function PlaceList({ places, onClick }: PlaceListProps) {
-  const theme = useTheme();
-  return (
-    <List
-      dense
-      sx={{
-        bgcolor: 'background.paper',
-        marginLeft: theme.spacing(3),
-        marginRight: theme.spacing(3),
-      }}
-    >
-      {places.map((value, index) => (
-        <ListItem
-          key={`${value}-${index}`}
-          secondaryAction={
-            <IconButton edge="end" aria-label="delete" onClick={() => onClick(index)}>
-              <DeleteIcon />
-            </IconButton>
-          }
-        >
-          <ListItemIcon>
-            <PlaceOutlined />
-          </ListItemIcon>
-          <ListItemText primary={`Place Name:   ${value}`} />
-        </ListItem>
-      ))}
-    </List>
-  );
-}
-
-interface PatrolTaskFormProps {
-  taskDesc: PatrolTaskDescription;
-  patrolWaypoints: string[];
-  onChange(patrolTaskDescription: PatrolTaskDescription): void;
-  allowSubmit(allow: boolean): void;
-}
-
-function PatrolTaskForm({ taskDesc, patrolWaypoints, onChange, allowSubmit }: PatrolTaskFormProps) {
-  const theme = useTheme();
-  const isScreenHeightLessThan800 = useMediaQuery('(max-height:800px)');
-  const onInputChange = (desc: PatrolTaskDescription) => {
-    allowSubmit(isPatrolTaskDescriptionValid(desc));
-    onChange(desc);
-  };
-  allowSubmit(isPatrolTaskDescriptionValid(taskDesc));
-
-  return (
-    <Grid container spacing={theme.spacing(2)} justifyContent="center" alignItems="center">
-      <Grid item xs={isScreenHeightLessThan800 ? 8 : 10}>
-        <Autocomplete
-          id="place-input"
-          freeSolo
-          fullWidth
-          options={patrolWaypoints.sort()}
-          onChange={(_ev, newValue) =>
-            newValue !== null &&
-            onInputChange({
-              ...taskDesc,
-              places: taskDesc.places.concat(newValue).filter((el: string) => el),
-            })
-          }
-          sx={{
-            '& .MuiOutlinedInput-root': {
-              height: isScreenHeightLessThan800 ? '3rem' : '3.5rem',
-              fontSize: isScreenHeightLessThan800 ? 14 : 20,
-            },
-          }}
-          renderInput={(params) => (
-            <TextField
-              {...params}
-              label="Place Name"
-              required={true}
-              InputLabelProps={{ style: { fontSize: isScreenHeightLessThan800 ? 14 : 20 } }}
-            />
-          )}
-        />
-      </Grid>
-      <Grid item xs={isScreenHeightLessThan800 ? 4 : 2}>
-        <PositiveIntField
-          id="loops"
-          label="Loops"
-          sx={{
-            '& .MuiOutlinedInput-root': {
-              height: isScreenHeightLessThan800 ? '3rem' : '3.5rem',
-              fontSize: isScreenHeightLessThan800 ? 14 : 20,
-            },
-          }}
-          value={taskDesc.rounds}
-          onChange={(_ev, val) => {
-            onInputChange({
-              ...taskDesc,
-              rounds: val,
-            });
-          }}
-        />
-      </Grid>
-      <Grid item xs={10}>
-        <PlaceList
-          places={taskDesc && taskDesc.places ? taskDesc.places : []}
-          onClick={(places_index) =>
-            taskDesc.places.splice(places_index, 1) &&
-            onInputChange({
-              ...taskDesc,
-            })
-          }
-        />
-      </Grid>
-    </Grid>
-  );
-}
 
 interface CustomComposeTaskFormProps {
   taskDesc: CustomComposeTaskDescription;
@@ -333,13 +196,6 @@ function FavoriteTask({
       </ListItem>
     </>
   );
-}
-
-export function defaultPatrolTask(): PatrolTaskDescription {
-  return {
-    places: [],
-    rounds: 1,
-  };
 }
 
 function defaultTaskDescription(taskCategory: string): TaskDescription | undefined {
